feat: 1.Add: popup-starter-validator module
