+++ popup-bom/popup-parent/popup-starter/popup-starter-autoconfigure/pom.xml
@@ -29,14 +29,10 @@
             <artifactId>popup-starter-logging</artifactId>
         </dependency>
 
+        <!-- validator -->
         <dependency>
             <groupId>com.photowey</groupId>
-            <artifactId>popup-exception-core</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-validation</artifactId>
+            <artifactId>popup-starter-validator</artifactId>
         </dependency>
 
     </dependencies>
